feat: organize contracts by file and version (#160)
