--- src/abi/index.ts
@@ -1,5 +1,0 @@
-import { rollupAdminLogicABI } from './rollupAdminLogicABI';
-import { sequencerInboxABI } from './sequencerInboxABI';
-import { safeL2ABI } from './safeL2ABI';
-
-export { rollupAdminLogicABI, sequencerInboxABI, safeL2ABI };
