+++ src/arbAggregatorPrepareTransactionRequest.ts
@@ -7,11 +7,11 @@ import {
   Transport,
 } from 'viem';
 
-import { arbAggregator } from './contracts';
+import { arbAggregatorABI, arbAggregatorAddress } from './contracts/ArbAggregator';
 import { upgradeExecutorEncodeFunctionData } from './upgradeExecutorEncodeFunctionData';
 import { GetFunctionName } from './types/utils';
 
-type ArbAggregatorAbi = typeof arbAggregator.abi;
+type ArbAggregatorAbi = typeof arbAggregatorABI;
 export type ArbAggregatorPrepareTransactionRequestFunctionName = GetFunctionName<ArbAggregatorAbi>;
 export type ArbAggregatorEncodeFunctionDataParameters<
   TFunctionName extends ArbAggregatorPrepareTransactionRequestFunctionName,
@@ -40,7 +40,7 @@ function arbAggregatorPrepareFunctionData<
 
   if (!upgradeExecutor) {
     return {
-      to: arbAggregator.address,
+      to: arbAggregatorAddress,
       data: arbAggregatorEncodeFunctionData(
         params as ArbAggregatorEncodeFunctionDataParameters<TFunctionName>,
       ),
@@ -53,7 +53,7 @@ function arbAggregatorPrepareFunctionData<
     data: upgradeExecutorEncodeFunctionData({
       functionName: 'executeCall',
       args: [
-        arbAggregator.address, // target
+        arbAggregatorAddress, // target
         arbAggregatorEncodeFunctionData(
           params as ArbAggregatorEncodeFunctionDataParameters<TFunctionName>,
         ), // targetCallData
@@ -82,7 +82,7 @@ export async function arbAggregatorPrepareTransactionRequest<
   // params is extending ArbAggregatorPrepareFunctionDataParameters, it's safe to cast
   const { to, data, value } = arbAggregatorPrepareFunctionData({
     ...params,
-    abi: arbAggregator.abi,
+    abi: arbAggregatorABI,
   } as unknown as ArbAggregatorPrepareFunctionDataParameters<TFunctionName>);
 
   // @ts-ignore (todo: fix viem type issue)
