--- src/contracts.ts
@@ -1,57 +1,0 @@
-import { parseAbi } from 'viem';
-
-import {
-  erc20ABI,
-  arbOwnerConfig,
-  arbOwnerPublicConfig,
-  rollupCreatorConfig,
-  tokenBridgeCreatorConfig,
-  arbGasInfoConfig,
-  arbAggregatorConfig,
-} from './generated';
-import { sequencerInboxABI, rollupAdminLogicABI } from './abi';
-
-export const erc20 = {
-  abi: erc20ABI,
-};
-
-export const arbOwner = {
-  ...arbOwnerConfig,
-  address: Object.values(arbOwnerConfig.address)[0],
-} as const;
-
-export const arbGasInfo = {
-  ...arbGasInfoConfig,
-  address: Object.values(arbGasInfoConfig.address)[0],
-} as const;
-
-export const arbOwnerPublic = {
-  ...arbOwnerPublicConfig,
-  address: Object.values(arbOwnerPublicConfig.address)[0],
-} as const;
-
-export const arbAggregator = {
-  ...arbAggregatorConfig,
-  address: Object.values(arbAggregatorConfig.address)[0],
-} as const;
-
-export const rollupCreator = rollupCreatorConfig;
-export const tokenBridgeCreator = tokenBridgeCreatorConfig;
-
-export const upgradeExecutor = {
-  abi: parseAbi([
-    'function execute(address upgrade, bytes upgradeCallData)',
-    'function executeCall(address target, bytes targetCallData)',
-    'function hasRole(bytes32 role, address account) public view returns (bool)',
-    'function grantRole(bytes32 role, address account)',
-    'function revokeRole(bytes32 role, address account)',
-  ]),
-};
-
-export const sequencerInbox = {
-  abi: sequencerInboxABI,
-};
-
-export const rollupAdminLogic = {
-  abi: rollupAdminLogicABI,
-};
