+++ src/contracts/RollupCreator/index.ts
@@ -1,0 +1,2 @@
+// export the latest version
+export * from './v1.1.0';
