+++ src/contracts/RollupCreator/v1.1.0.ts
@@ -1,0 +1,242 @@
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// RollupCreator
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
+ * -
+ * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
+ * -
+ * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
+ * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
+ * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
+ */
+export const rollupCreatorABI = [
+  { stateMutability: 'nonpayable', type: 'constructor', inputs: [] },
+  {
+    type: 'event',
+    anonymous: false,
+    inputs: [
+      { name: 'previousOwner', internalType: 'address', type: 'address', indexed: true },
+      { name: 'newOwner', internalType: 'address', type: 'address', indexed: true },
+    ],
+    name: 'OwnershipTransferred',
+  },
+  {
+    type: 'event',
+    anonymous: false,
+    inputs: [
+      { name: 'rollupAddress', internalType: 'address', type: 'address', indexed: true },
+      { name: 'nativeToken', internalType: 'address', type: 'address', indexed: true },
+      { name: 'inboxAddress', internalType: 'address', type: 'address', indexed: false },
+      { name: 'outbox', internalType: 'address', type: 'address', indexed: false },
+      { name: 'rollupEventInbox', internalType: 'address', type: 'address', indexed: false },
+      { name: 'challengeManager', internalType: 'address', type: 'address', indexed: false },
+      { name: 'adminProxy', internalType: 'address', type: 'address', indexed: false },
+      { name: 'sequencerInbox', internalType: 'address', type: 'address', indexed: false },
+      { name: 'bridge', internalType: 'address', type: 'address', indexed: false },
+      { name: 'upgradeExecutor', internalType: 'address', type: 'address', indexed: false },
+      { name: 'validatorUtils', internalType: 'address', type: 'address', indexed: false },
+      { name: 'validatorWalletCreator', internalType: 'address', type: 'address', indexed: false },
+    ],
+    name: 'RollupCreated',
+  },
+  { type: 'event', anonymous: false, inputs: [], name: 'TemplatesUpdated' },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'bridgeCreator',
+    outputs: [{ name: '', internalType: 'contract BridgeCreator', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'challengeManagerTemplate',
+    outputs: [{ name: '', internalType: 'contract IChallengeManager', type: 'address' }],
+  },
+  {
+    stateMutability: 'payable',
+    type: 'function',
+    inputs: [
+      {
+        name: 'deployParams',
+        internalType: 'struct RollupCreator.RollupDeploymentParams',
+        type: 'tuple',
+        components: [
+          {
+            name: 'config',
+            internalType: 'struct Config',
+            type: 'tuple',
+            components: [
+              { name: 'confirmPeriodBlocks', internalType: 'uint64', type: 'uint64' },
+              { name: 'extraChallengeTimeBlocks', internalType: 'uint64', type: 'uint64' },
+              { name: 'stakeToken', internalType: 'address', type: 'address' },
+              { name: 'baseStake', internalType: 'uint256', type: 'uint256' },
+              { name: 'wasmModuleRoot', internalType: 'bytes32', type: 'bytes32' },
+              { name: 'owner', internalType: 'address', type: 'address' },
+              { name: 'loserStakeEscrow', internalType: 'address', type: 'address' },
+              { name: 'chainId', internalType: 'uint256', type: 'uint256' },
+              { name: 'chainConfig', internalType: 'string', type: 'string' },
+              { name: 'genesisBlockNum', internalType: 'uint64', type: 'uint64' },
+              {
+                name: 'sequencerInboxMaxTimeVariation',
+                internalType: 'struct ISequencerInbox.MaxTimeVariation',
+                type: 'tuple',
+                components: [
+                  { name: 'delayBlocks', internalType: 'uint256', type: 'uint256' },
+                  { name: 'futureBlocks', internalType: 'uint256', type: 'uint256' },
+                  { name: 'delaySeconds', internalType: 'uint256', type: 'uint256' },
+                  { name: 'futureSeconds', internalType: 'uint256', type: 'uint256' },
+                ],
+              },
+            ],
+          },
+          { name: 'batchPoster', internalType: 'address', type: 'address' },
+          { name: 'validators', internalType: 'address[]', type: 'address[]' },
+          { name: 'maxDataSize', internalType: 'uint256', type: 'uint256' },
+          { name: 'nativeToken', internalType: 'address', type: 'address' },
+          { name: 'deployFactoriesToL2', internalType: 'bool', type: 'bool' },
+          { name: 'maxFeePerGasForRetryables', internalType: 'uint256', type: 'uint256' },
+        ],
+      },
+    ],
+    name: 'createRollup',
+    outputs: [{ name: '', internalType: 'address', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'l2FactoriesDeployer',
+    outputs: [{ name: '', internalType: 'contract DeployHelper', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'osp',
+    outputs: [{ name: '', internalType: 'contract IOneStepProofEntry', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'owner',
+    outputs: [{ name: '', internalType: 'address', type: 'address' }],
+  },
+  {
+    stateMutability: 'nonpayable',
+    type: 'function',
+    inputs: [],
+    name: 'renounceOwnership',
+    outputs: [],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'rollupAdminLogic',
+    outputs: [{ name: '', internalType: 'contract IRollupAdmin', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'rollupUserLogic',
+    outputs: [{ name: '', internalType: 'contract IRollupUser', type: 'address' }],
+  },
+  {
+    stateMutability: 'nonpayable',
+    type: 'function',
+    inputs: [
+      { name: '_bridgeCreator', internalType: 'contract BridgeCreator', type: 'address' },
+      { name: '_osp', internalType: 'contract IOneStepProofEntry', type: 'address' },
+      {
+        name: '_challengeManagerLogic',
+        internalType: 'contract IChallengeManager',
+        type: 'address',
+      },
+      { name: '_rollupAdminLogic', internalType: 'contract IRollupAdmin', type: 'address' },
+      { name: '_rollupUserLogic', internalType: 'contract IRollupUser', type: 'address' },
+      { name: '_upgradeExecutorLogic', internalType: 'contract IUpgradeExecutor', type: 'address' },
+      { name: '_validatorUtils', internalType: 'address', type: 'address' },
+      { name: '_validatorWalletCreator', internalType: 'address', type: 'address' },
+      { name: '_l2FactoriesDeployer', internalType: 'contract DeployHelper', type: 'address' },
+    ],
+    name: 'setTemplates',
+    outputs: [],
+  },
+  {
+    stateMutability: 'nonpayable',
+    type: 'function',
+    inputs: [{ name: 'newOwner', internalType: 'address', type: 'address' }],
+    name: 'transferOwnership',
+    outputs: [],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'upgradeExecutorLogic',
+    outputs: [{ name: '', internalType: 'contract IUpgradeExecutor', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'validatorUtils',
+    outputs: [{ name: '', internalType: 'address', type: 'address' }],
+  },
+  {
+    stateMutability: 'view',
+    type: 'function',
+    inputs: [],
+    name: 'validatorWalletCreator',
+    outputs: [{ name: '', internalType: 'address', type: 'address' }],
+  },
+  { stateMutability: 'payable', type: 'receive' },
+] as const;
+
+/**
+ * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
+ * -
+ * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
+ * -
+ * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
+ * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
+ * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
+ */
+export const rollupCreatorAddress = {
+  1: '0x90D68B056c411015eaE3EC0b98AD94E2C91419F1',
+  1337: '0x596eAbE0291D4cdAfAC7ef53D16C92Bf6922b5e0',
+  8453: '0x850F050C65B34966895AdA26a4D06923901916DB',
+  17000: '0xB512078282F462Ba104231ad856464Ceb0a7747e',
+  42161: '0x9CAd81628aB7D8e239F1A5B497313341578c5F71',
+  42170: '0x9CAd81628aB7D8e239F1A5B497313341578c5F71',
+  84532: '0x1E0921818df948c338380e722C8aE91Bb285763C',
+  412346: '0x3BaF9f08bAD68869eEdEa90F2Cc546Bd80F1A651',
+  421614: '0x06E341073b2749e0Bb9912461351f716DeCDa9b0',
+  11155111: '0xfBD0B034e6305788007f6e0123cc5EaE701a5751',
+} as const;
+
+/**
+ * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
+ * -
+ * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
+ * -
+ * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
+ * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
+ * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
+ * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
+ */
+export const rollupCreatorConfig = {
+  address: rollupCreatorAddress,
+  abi: rollupCreatorABI,
+} as const;
