--- src/generated.ts
@@ -1,1413 +1,0 @@
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// ArbAggregator
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006d)
- */
-export const arbAggregatorABI = [
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newBatchPoster', internalType: 'address', type: 'address' }],
-    name: 'addBatchPoster',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getBatchPosters',
-    outputs: [{ name: '', internalType: 'address[]', type: 'address[]' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getDefaultAggregator',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'batchPoster', internalType: 'address', type: 'address' }],
-    name: 'getFeeCollector',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'addr', internalType: 'address', type: 'address' }],
-    name: 'getPreferredAggregator',
-    outputs: [
-      { name: '', internalType: 'address', type: 'address' },
-      { name: '', internalType: 'bool', type: 'bool' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'aggregator', internalType: 'address', type: 'address' }],
-    name: 'getTxBaseFee',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'batchPoster', internalType: 'address', type: 'address' },
-      { name: 'newFeeCollector', internalType: 'address', type: 'address' },
-    ],
-    name: 'setFeeCollector',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'aggregator', internalType: 'address', type: 'address' },
-      { name: 'feeInL1Gas', internalType: 'uint256', type: 'uint256' },
-    ],
-    name: 'setTxBaseFee',
-    outputs: [],
-  },
-] as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006d)
- */
-export const arbAggregatorAddress = {
-  421614: '0x000000000000000000000000000000000000006D',
-} as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006d)
- */
-export const arbAggregatorConfig = {
-  address: arbAggregatorAddress,
-  abi: arbAggregatorABI,
-} as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// ArbGasInfo
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006c)
- */
-export const arbGasInfoABI = [
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getAmortizedCostCapBips',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getCurrentTxL1GasFees',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getGasAccountingParams',
-    outputs: [
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getGasBacklog',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getGasBacklogTolerance',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1BaseFeeEstimate',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1BaseFeeEstimateInertia',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1FeesAvailable',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1GasPriceEstimate',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1PricingEquilibrationUnits',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1PricingFundsDueForRewards',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1PricingSurplus',
-    outputs: [{ name: '', internalType: 'int256', type: 'int256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1PricingUnitsSinceUpdate',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1RewardRate',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getL1RewardRecipient',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getLastL1PricingSurplus',
-    outputs: [{ name: '', internalType: 'int256', type: 'int256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getLastL1PricingUpdateTime',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getMinimumGasPrice',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getPerBatchGasCharge',
-    outputs: [{ name: '', internalType: 'int64', type: 'int64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getPricesInArbGas',
-    outputs: [
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'aggregator', internalType: 'address', type: 'address' }],
-    name: 'getPricesInArbGasWithAggregator',
-    outputs: [
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getPricesInWei',
-    outputs: [
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'aggregator', internalType: 'address', type: 'address' }],
-    name: 'getPricesInWeiWithAggregator',
-    outputs: [
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-      { name: '', internalType: 'uint256', type: 'uint256' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getPricingInertia',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-] as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006c)
- */
-export const arbGasInfoAddress = {
-  421614: '0x000000000000000000000000000000000000006C',
-} as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006c)
- */
-export const arbGasInfoConfig = { address: arbGasInfoAddress, abi: arbGasInfoABI } as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// ArbOwner
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x0000000000000000000000000000000000000070)
- */
-export const arbOwnerABI = [
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'method', internalType: 'bytes4', type: 'bytes4', indexed: true },
-      { name: 'owner', internalType: 'address', type: 'address', indexed: true },
-      { name: 'data', internalType: 'bytes', type: 'bytes', indexed: false },
-    ],
-    name: 'OwnerActs',
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newOwner', internalType: 'address', type: 'address' }],
-    name: 'addChainOwner',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getAllChainOwners',
-    outputs: [{ name: '', internalType: 'address[]', type: 'address[]' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getInfraFeeAccount',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getNetworkFeeAccount',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'addr', internalType: 'address', type: 'address' }],
-    name: 'isChainOwner',
-    outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'maxWeiToRelease', internalType: 'uint256', type: 'uint256' }],
-    name: 'releaseL1PricerSurplusFunds',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'ownerToRemove', internalType: 'address', type: 'address' }],
-    name: 'removeChainOwner',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'newVersion', internalType: 'uint64', type: 'uint64' },
-      { name: 'timestamp', internalType: 'uint64', type: 'uint64' },
-    ],
-    name: 'scheduleArbOSUpgrade',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'cap', internalType: 'uint64', type: 'uint64' }],
-    name: 'setAmortizedCostCapBips',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'level', internalType: 'uint64', type: 'uint64' }],
-    name: 'setBrotliCompressionLevel',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'chainConfig', internalType: 'string', type: 'string' }],
-    name: 'setChainConfig',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newInfraFeeAccount', internalType: 'address', type: 'address' }],
-    name: 'setInfraFeeAccount',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'inertia', internalType: 'uint64', type: 'uint64' }],
-    name: 'setL1BaseFeeEstimateInertia',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'pricePerUnit', internalType: 'uint256', type: 'uint256' }],
-    name: 'setL1PricePerUnit',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'equilibrationUnits', internalType: 'uint256', type: 'uint256' }],
-    name: 'setL1PricingEquilibrationUnits',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'inertia', internalType: 'uint64', type: 'uint64' }],
-    name: 'setL1PricingInertia',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'weiPerUnit', internalType: 'uint64', type: 'uint64' }],
-    name: 'setL1PricingRewardRate',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'recipient', internalType: 'address', type: 'address' }],
-    name: 'setL1PricingRewardRecipient',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'priceInWei', internalType: 'uint256', type: 'uint256' }],
-    name: 'setL2BaseFee',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'sec', internalType: 'uint64', type: 'uint64' }],
-    name: 'setL2GasBacklogTolerance',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'sec', internalType: 'uint64', type: 'uint64' }],
-    name: 'setL2GasPricingInertia',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'limit', internalType: 'uint64', type: 'uint64' }],
-    name: 'setMaxTxGasLimit',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'priceInWei', internalType: 'uint256', type: 'uint256' }],
-    name: 'setMinimumL2BaseFee',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newNetworkFeeAccount', internalType: 'address', type: 'address' }],
-    name: 'setNetworkFeeAccount',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'cost', internalType: 'int64', type: 'int64' }],
-    name: 'setPerBatchGasCharge',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'limit', internalType: 'uint64', type: 'uint64' }],
-    name: 'setSpeedLimit',
-    outputs: [],
-  },
-] as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x0000000000000000000000000000000000000070)
- */
-export const arbOwnerAddress = {
-  421614: '0x0000000000000000000000000000000000000070',
-} as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x0000000000000000000000000000000000000070)
- */
-export const arbOwnerConfig = { address: arbOwnerAddress, abi: arbOwnerABI } as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// ArbOwnerPublic
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006b)
- */
-export const arbOwnerPublicABI = [
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [{ name: 'rectifiedOwner', internalType: 'address', type: 'address', indexed: false }],
-    name: 'ChainOwnerRectified',
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getAllChainOwners',
-    outputs: [{ name: '', internalType: 'address[]', type: 'address[]' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getBrotliCompressionLevel',
-    outputs: [{ name: '', internalType: 'uint64', type: 'uint64' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getInfraFeeAccount',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getNetworkFeeAccount',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'getScheduledUpgrade',
-    outputs: [
-      { name: 'arbosVersion', internalType: 'uint64', type: 'uint64' },
-      { name: 'scheduledForTimestamp', internalType: 'uint64', type: 'uint64' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'addr', internalType: 'address', type: 'address' }],
-    name: 'isChainOwner',
-    outputs: [{ name: '', internalType: 'bool', type: 'bool' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'ownerToRectify', internalType: 'address', type: 'address' }],
-    name: 'rectifyChainOwner',
-    outputs: [],
-  },
-] as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006b)
- */
-export const arbOwnerPublicAddress = {
-  421614: '0x000000000000000000000000000000000000006b',
-} as const;
-
-/**
- * [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x000000000000000000000000000000000000006b)
- */
-export const arbOwnerPublicConfig = {
-  address: arbOwnerPublicAddress,
-  abi: arbOwnerPublicABI,
-} as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// ERC20
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-export const erc20ABI = [
-  {
-    type: 'event',
-    inputs: [
-      { name: 'owner', type: 'address', indexed: true },
-      { name: 'spender', type: 'address', indexed: true },
-      { name: 'value', type: 'uint256', indexed: false },
-    ],
-    name: 'Approval',
-  },
-  {
-    type: 'event',
-    inputs: [
-      { name: 'from', type: 'address', indexed: true },
-      { name: 'to', type: 'address', indexed: true },
-      { name: 'value', type: 'uint256', indexed: false },
-    ],
-    name: 'Transfer',
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [
-      { name: 'owner', type: 'address' },
-      { name: 'spender', type: 'address' },
-    ],
-    name: 'allowance',
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'spender', type: 'address' },
-      { name: 'amount', type: 'uint256' },
-    ],
-    name: 'approve',
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'account', type: 'address' }],
-    name: 'balanceOf',
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'decimals',
-    outputs: [{ type: 'uint8' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'name',
-    outputs: [{ type: 'string' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'symbol',
-    outputs: [{ type: 'string' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'totalSupply',
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'recipient', type: 'address' },
-      { name: 'amount', type: 'uint256' },
-    ],
-    name: 'transfer',
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'sender', type: 'address' },
-      { name: 'recipient', type: 'address' },
-      { name: 'amount', type: 'uint256' },
-    ],
-    name: 'transferFrom',
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'spender', type: 'address' },
-      { name: 'addedValue', type: 'uint256' },
-    ],
-    name: 'increaseAllowance',
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'spender', type: 'address' },
-      { name: 'subtractedValue', type: 'uint256' },
-    ],
-    name: 'decreaseAllowance',
-    outputs: [{ type: 'bool' }],
-  },
-] as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// RollupCreator
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
- */
-export const rollupCreatorABI = [
-  { stateMutability: 'nonpayable', type: 'constructor', inputs: [] },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'previousOwner', internalType: 'address', type: 'address', indexed: true },
-      { name: 'newOwner', internalType: 'address', type: 'address', indexed: true },
-    ],
-    name: 'OwnershipTransferred',
-  },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'rollupAddress', internalType: 'address', type: 'address', indexed: true },
-      { name: 'nativeToken', internalType: 'address', type: 'address', indexed: true },
-      { name: 'inboxAddress', internalType: 'address', type: 'address', indexed: false },
-      { name: 'outbox', internalType: 'address', type: 'address', indexed: false },
-      { name: 'rollupEventInbox', internalType: 'address', type: 'address', indexed: false },
-      { name: 'challengeManager', internalType: 'address', type: 'address', indexed: false },
-      { name: 'adminProxy', internalType: 'address', type: 'address', indexed: false },
-      { name: 'sequencerInbox', internalType: 'address', type: 'address', indexed: false },
-      { name: 'bridge', internalType: 'address', type: 'address', indexed: false },
-      { name: 'upgradeExecutor', internalType: 'address', type: 'address', indexed: false },
-      { name: 'validatorUtils', internalType: 'address', type: 'address', indexed: false },
-      { name: 'validatorWalletCreator', internalType: 'address', type: 'address', indexed: false },
-    ],
-    name: 'RollupCreated',
-  },
-  { type: 'event', anonymous: false, inputs: [], name: 'TemplatesUpdated' },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'bridgeCreator',
-    outputs: [{ name: '', internalType: 'contract BridgeCreator', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'challengeManagerTemplate',
-    outputs: [{ name: '', internalType: 'contract IChallengeManager', type: 'address' }],
-  },
-  {
-    stateMutability: 'payable',
-    type: 'function',
-    inputs: [
-      {
-        name: 'deployParams',
-        internalType: 'struct RollupCreator.RollupDeploymentParams',
-        type: 'tuple',
-        components: [
-          {
-            name: 'config',
-            internalType: 'struct Config',
-            type: 'tuple',
-            components: [
-              { name: 'confirmPeriodBlocks', internalType: 'uint64', type: 'uint64' },
-              { name: 'extraChallengeTimeBlocks', internalType: 'uint64', type: 'uint64' },
-              { name: 'stakeToken', internalType: 'address', type: 'address' },
-              { name: 'baseStake', internalType: 'uint256', type: 'uint256' },
-              { name: 'wasmModuleRoot', internalType: 'bytes32', type: 'bytes32' },
-              { name: 'owner', internalType: 'address', type: 'address' },
-              { name: 'loserStakeEscrow', internalType: 'address', type: 'address' },
-              { name: 'chainId', internalType: 'uint256', type: 'uint256' },
-              { name: 'chainConfig', internalType: 'string', type: 'string' },
-              { name: 'genesisBlockNum', internalType: 'uint64', type: 'uint64' },
-              {
-                name: 'sequencerInboxMaxTimeVariation',
-                internalType: 'struct ISequencerInbox.MaxTimeVariation',
-                type: 'tuple',
-                components: [
-                  { name: 'delayBlocks', internalType: 'uint256', type: 'uint256' },
-                  { name: 'futureBlocks', internalType: 'uint256', type: 'uint256' },
-                  { name: 'delaySeconds', internalType: 'uint256', type: 'uint256' },
-                  { name: 'futureSeconds', internalType: 'uint256', type: 'uint256' },
-                ],
-              },
-            ],
-          },
-          { name: 'batchPoster', internalType: 'address', type: 'address' },
-          { name: 'validators', internalType: 'address[]', type: 'address[]' },
-          { name: 'maxDataSize', internalType: 'uint256', type: 'uint256' },
-          { name: 'nativeToken', internalType: 'address', type: 'address' },
-          { name: 'deployFactoriesToL2', internalType: 'bool', type: 'bool' },
-          { name: 'maxFeePerGasForRetryables', internalType: 'uint256', type: 'uint256' },
-        ],
-      },
-    ],
-    name: 'createRollup',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2FactoriesDeployer',
-    outputs: [{ name: '', internalType: 'contract DeployHelper', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'osp',
-    outputs: [{ name: '', internalType: 'contract IOneStepProofEntry', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'owner',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [],
-    name: 'renounceOwnership',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'rollupAdminLogic',
-    outputs: [{ name: '', internalType: 'contract IRollupAdmin', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'rollupUserLogic',
-    outputs: [{ name: '', internalType: 'contract IRollupUser', type: 'address' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: '_bridgeCreator', internalType: 'contract BridgeCreator', type: 'address' },
-      { name: '_osp', internalType: 'contract IOneStepProofEntry', type: 'address' },
-      {
-        name: '_challengeManagerLogic',
-        internalType: 'contract IChallengeManager',
-        type: 'address',
-      },
-      { name: '_rollupAdminLogic', internalType: 'contract IRollupAdmin', type: 'address' },
-      { name: '_rollupUserLogic', internalType: 'contract IRollupUser', type: 'address' },
-      { name: '_upgradeExecutorLogic', internalType: 'contract IUpgradeExecutor', type: 'address' },
-      { name: '_validatorUtils', internalType: 'address', type: 'address' },
-      { name: '_validatorWalletCreator', internalType: 'address', type: 'address' },
-      { name: '_l2FactoriesDeployer', internalType: 'contract DeployHelper', type: 'address' },
-    ],
-    name: 'setTemplates',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newOwner', internalType: 'address', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'upgradeExecutorLogic',
-    outputs: [{ name: '', internalType: 'contract IUpgradeExecutor', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'validatorUtils',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'validatorWalletCreator',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  { stateMutability: 'payable', type: 'receive' },
-] as const;
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
- */
-export const rollupCreatorAddress = {
-  1: '0x90D68B056c411015eaE3EC0b98AD94E2C91419F1',
-  1337: '0x596eAbE0291D4cdAfAC7ef53D16C92Bf6922b5e0',
-  8453: '0x850F050C65B34966895AdA26a4D06923901916DB',
-  17000: '0xB512078282F462Ba104231ad856464Ceb0a7747e',
-  42161: '0x9CAd81628aB7D8e239F1A5B497313341578c5F71',
-  42170: '0x9CAd81628aB7D8e239F1A5B497313341578c5F71',
-  84532: '0x1E0921818df948c338380e722C8aE91Bb285763C',
-  412346: '0x3BaF9f08bAD68869eEdEa90F2Cc546Bd80F1A651',
-  421614: '0x06E341073b2749e0Bb9912461351f716DeCDa9b0',
-  11155111: '0xfBD0B034e6305788007f6e0123cc5EaE701a5751',
-} as const;
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x90d68b056c411015eae3ec0b98ad94e2c91419f1)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x850F050C65B34966895AdA26a4D06923901916DB)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x9CAd81628aB7D8e239F1A5B497313341578c5F71)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0x1E0921818df948c338380e722C8aE91Bb285763C)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x06E341073b2749e0Bb9912461351f716DeCDa9b0)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0xfbd0b034e6305788007f6e0123cc5eae701a5751)
- */
-export const rollupCreatorConfig = {
-  address: rollupCreatorAddress,
-  abi: rollupCreatorABI,
-} as const;
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// TokenBridgeCreator
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x60D9A46F24D5a35b95A78Dd3E793e55D94EE0660)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x4C240987d6fE4fa8C7a0004986e3db563150CA55)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x2f5624dc8800dfA0A82AC03509Ef8bb8E7Ac000e)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x8B9D9490a68B1F16ac8A21DdAE5Fd7aB9d708c14)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0xFC71d21a4FE10Cc0d34745ba9c713836f82f8DE3)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x56C486D3786fA26cc61473C499A36Eb9CC1FbD8E)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0x7edb2dfBeEf9417e0454A80c51EE0C034e45a570)
- */
-export const tokenBridgeCreatorABI = [
-  { stateMutability: 'nonpayable', type: 'constructor', inputs: [] },
-  { type: 'error', inputs: [], name: 'L1AtomicTokenBridgeCreator_L2FactoryCannotBeChanged' },
-  { type: 'error', inputs: [], name: 'L1AtomicTokenBridgeCreator_OnlyRollupOwner' },
-  { type: 'error', inputs: [], name: 'L1AtomicTokenBridgeCreator_ProxyAdminNotFound' },
-  { type: 'error', inputs: [], name: 'L1AtomicTokenBridgeCreator_RollupOwnershipMisconfig' },
-  { type: 'error', inputs: [], name: 'L1AtomicTokenBridgeCreator_TemplatesNotSet' },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [{ name: 'version', internalType: 'uint8', type: 'uint8', indexed: false }],
-    name: 'Initialized',
-  },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'inbox', internalType: 'address', type: 'address', indexed: true },
-      { name: 'owner', internalType: 'address', type: 'address', indexed: true },
-      {
-        name: 'l1Deployment',
-        internalType: 'struct L1DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-        ],
-        indexed: false,
-      },
-      {
-        name: 'l2Deployment',
-        internalType: 'struct L2DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-          { name: 'proxyAdmin', internalType: 'address', type: 'address' },
-          { name: 'beaconProxyFactory', internalType: 'address', type: 'address' },
-          { name: 'upgradeExecutor', internalType: 'address', type: 'address' },
-          { name: 'multicall', internalType: 'address', type: 'address' },
-        ],
-        indexed: false,
-      },
-      { name: 'proxyAdmin', internalType: 'address', type: 'address', indexed: false },
-      { name: 'upgradeExecutor', internalType: 'address', type: 'address', indexed: false },
-    ],
-    name: 'OrbitTokenBridgeCreated',
-  },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'inbox', internalType: 'address', type: 'address', indexed: true },
-      {
-        name: 'l1',
-        internalType: 'struct L1DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-        ],
-        indexed: false,
-      },
-      {
-        name: 'l2',
-        internalType: 'struct L2DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-          { name: 'proxyAdmin', internalType: 'address', type: 'address' },
-          { name: 'beaconProxyFactory', internalType: 'address', type: 'address' },
-          { name: 'upgradeExecutor', internalType: 'address', type: 'address' },
-          { name: 'multicall', internalType: 'address', type: 'address' },
-        ],
-        indexed: false,
-      },
-    ],
-    name: 'OrbitTokenBridgeDeploymentSet',
-  },
-  { type: 'event', anonymous: false, inputs: [], name: 'OrbitTokenBridgeTemplatesUpdated' },
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'previousOwner', internalType: 'address', type: 'address', indexed: true },
-      { name: 'newOwner', internalType: 'address', type: 'address', indexed: true },
-    ],
-    name: 'OwnershipTransferred',
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'canonicalL2FactoryAddress',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'payable',
-    type: 'function',
-    inputs: [
-      { name: 'inbox', internalType: 'address', type: 'address' },
-      { name: 'rollupOwner', internalType: 'address', type: 'address' },
-      { name: 'maxGasForContracts', internalType: 'uint256', type: 'uint256' },
-      { name: 'gasPriceBid', internalType: 'uint256', type: 'uint256' },
-    ],
-    name: 'createTokenBridge',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'gasLimitForL2FactoryDeployment',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: 'inbox', internalType: 'address', type: 'address' }],
-    name: 'getRouter',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: '', internalType: 'address', type: 'address' }],
-    name: 'inboxToL1Deployment',
-    outputs: [
-      { name: 'router', internalType: 'address', type: 'address' },
-      { name: 'standardGateway', internalType: 'address', type: 'address' },
-      { name: 'customGateway', internalType: 'address', type: 'address' },
-      { name: 'wethGateway', internalType: 'address', type: 'address' },
-      { name: 'weth', internalType: 'address', type: 'address' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [{ name: '', internalType: 'address', type: 'address' }],
-    name: 'inboxToL2Deployment',
-    outputs: [
-      { name: 'router', internalType: 'address', type: 'address' },
-      { name: 'standardGateway', internalType: 'address', type: 'address' },
-      { name: 'customGateway', internalType: 'address', type: 'address' },
-      { name: 'wethGateway', internalType: 'address', type: 'address' },
-      { name: 'weth', internalType: 'address', type: 'address' },
-      { name: 'proxyAdmin', internalType: 'address', type: 'address' },
-      { name: 'beaconProxyFactory', internalType: 'address', type: 'address' },
-      { name: 'upgradeExecutor', internalType: 'address', type: 'address' },
-      { name: 'multicall', internalType: 'address', type: 'address' },
-    ],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      {
-        name: '_retryableSender',
-        internalType: 'contract L1TokenBridgeRetryableSender',
-        type: 'address',
-      },
-    ],
-    name: 'initialize',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l1Multicall',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l1Templates',
-    outputs: [
-      { name: 'routerTemplate', internalType: 'contract L1GatewayRouter', type: 'address' },
-      { name: 'standardGatewayTemplate', internalType: 'contract L1ERC20Gateway', type: 'address' },
-      { name: 'customGatewayTemplate', internalType: 'contract L1CustomGateway', type: 'address' },
-      { name: 'wethGatewayTemplate', internalType: 'contract L1WethGateway', type: 'address' },
-      {
-        name: 'feeTokenBasedRouterTemplate',
-        internalType: 'contract L1OrbitGatewayRouter',
-        type: 'address',
-      },
-      {
-        name: 'feeTokenBasedStandardGatewayTemplate',
-        internalType: 'contract L1OrbitERC20Gateway',
-        type: 'address',
-      },
-      {
-        name: 'feeTokenBasedCustomGatewayTemplate',
-        internalType: 'contract L1OrbitCustomGateway',
-        type: 'address',
-      },
-      { name: 'upgradeExecutor', internalType: 'contract IUpgradeExecutor', type: 'address' },
-    ],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l1Weth',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2CustomGatewayTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2MulticallTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2RouterTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2StandardGatewayTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2TokenBridgeFactoryTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2WethGatewayTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'l2WethTemplate',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'owner',
-    outputs: [{ name: '', internalType: 'address', type: 'address' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [],
-    name: 'renounceOwnership',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'retryableSender',
-    outputs: [{ name: '', internalType: 'contract L1TokenBridgeRetryableSender', type: 'address' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      { name: 'inbox', internalType: 'address', type: 'address' },
-      {
-        name: 'l1Deployment',
-        internalType: 'struct L1DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-        ],
-      },
-      {
-        name: 'l2Deployment',
-        internalType: 'struct L2DeploymentAddresses',
-        type: 'tuple',
-        components: [
-          { name: 'router', internalType: 'address', type: 'address' },
-          { name: 'standardGateway', internalType: 'address', type: 'address' },
-          { name: 'customGateway', internalType: 'address', type: 'address' },
-          { name: 'wethGateway', internalType: 'address', type: 'address' },
-          { name: 'weth', internalType: 'address', type: 'address' },
-          { name: 'proxyAdmin', internalType: 'address', type: 'address' },
-          { name: 'beaconProxyFactory', internalType: 'address', type: 'address' },
-          { name: 'upgradeExecutor', internalType: 'address', type: 'address' },
-          { name: 'multicall', internalType: 'address', type: 'address' },
-        ],
-      },
-    ],
-    name: 'setDeployment',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [
-      {
-        name: '_l1Templates',
-        internalType: 'struct L1AtomicTokenBridgeCreator.L1Templates',
-        type: 'tuple',
-        components: [
-          { name: 'routerTemplate', internalType: 'contract L1GatewayRouter', type: 'address' },
-          {
-            name: 'standardGatewayTemplate',
-            internalType: 'contract L1ERC20Gateway',
-            type: 'address',
-          },
-          {
-            name: 'customGatewayTemplate',
-            internalType: 'contract L1CustomGateway',
-            type: 'address',
-          },
-          { name: 'wethGatewayTemplate', internalType: 'contract L1WethGateway', type: 'address' },
-          {
-            name: 'feeTokenBasedRouterTemplate',
-            internalType: 'contract L1OrbitGatewayRouter',
-            type: 'address',
-          },
-          {
-            name: 'feeTokenBasedStandardGatewayTemplate',
-            internalType: 'contract L1OrbitERC20Gateway',
-            type: 'address',
-          },
-          {
-            name: 'feeTokenBasedCustomGatewayTemplate',
-            internalType: 'contract L1OrbitCustomGateway',
-            type: 'address',
-          },
-          { name: 'upgradeExecutor', internalType: 'contract IUpgradeExecutor', type: 'address' },
-        ],
-      },
-      { name: '_l2TokenBridgeFactoryTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2RouterTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2StandardGatewayTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2CustomGatewayTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2WethGatewayTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2WethTemplate', internalType: 'address', type: 'address' },
-      { name: '_l2MulticallTemplate', internalType: 'address', type: 'address' },
-      { name: '_l1Weth', internalType: 'address', type: 'address' },
-      { name: '_l1Multicall', internalType: 'address', type: 'address' },
-      { name: '_gasLimitForL2FactoryDeployment', internalType: 'uint256', type: 'uint256' },
-    ],
-    name: 'setTemplates',
-    outputs: [],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newOwner', internalType: 'address', type: 'address' }],
-    name: 'transferOwnership',
-    outputs: [],
-  },
-] as const;
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x60D9A46F24D5a35b95A78Dd3E793e55D94EE0660)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x4C240987d6fE4fa8C7a0004986e3db563150CA55)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x2f5624dc8800dfA0A82AC03509Ef8bb8E7Ac000e)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x8B9D9490a68B1F16ac8A21DdAE5Fd7aB9d708c14)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0xFC71d21a4FE10Cc0d34745ba9c713836f82f8DE3)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x56C486D3786fA26cc61473C499A36Eb9CC1FbD8E)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0x7edb2dfBeEf9417e0454A80c51EE0C034e45a570)
- */
-export const tokenBridgeCreatorAddress = {
-  1: '0x60D9A46F24D5a35b95A78Dd3E793e55D94EE0660',
-  1337: '0x54B4D4e578E10178a6cA602bdb6df0F213296Af4',
-  8453: '0x4C240987d6fE4fa8C7a0004986e3db563150CA55',
-  17000: '0xac890ED9bC2494C053cE701F138958df95966d94',
-  42161: '0x2f5624dc8800dfA0A82AC03509Ef8bb8E7Ac000e',
-  42170: '0x8B9D9490a68B1F16ac8A21DdAE5Fd7aB9d708c14',
-  84532: '0xFC71d21a4FE10Cc0d34745ba9c713836f82f8DE3',
-  412346: '0x38F35Af53bF913c439eaB06A367e09D6eb253492',
-  421614: '0x56C486D3786fA26cc61473C499A36Eb9CC1FbD8E',
-  11155111: '0x7edb2dfBeEf9417e0454A80c51EE0C034e45a570',
-} as const;
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x60D9A46F24D5a35b95A78Dd3E793e55D94EE0660)
- * -
- * - [__View Contract on Base Basescan__](https://basescan.org/address/0x4C240987d6fE4fa8C7a0004986e3db563150CA55)
- * -
- * - [__View Contract on Arbitrum One Arbiscan__](https://arbiscan.io/address/0x2f5624dc8800dfA0A82AC03509Ef8bb8E7Ac000e)
- * - [__View Contract on Arbitrum Nova Arbiscan__](https://nova.arbiscan.io/address/0x8B9D9490a68B1F16ac8A21DdAE5Fd7aB9d708c14)
- * - [__View Contract on Base Sepolia Blockscout__](https://base-sepolia.blockscout.com/address/0xFC71d21a4FE10Cc0d34745ba9c713836f82f8DE3)
- * - [__View Contract on Arbitrum Sepolia Blockscout__](https://sepolia-explorer.arbitrum.io/address/0x56C486D3786fA26cc61473C499A36Eb9CC1FbD8E)
- * - [__View Contract on Sepolia Etherscan__](https://sepolia.etherscan.io/address/0x7edb2dfBeEf9417e0454A80c51EE0C034e45a570)
- */
-export const tokenBridgeCreatorConfig = {
-  address: tokenBridgeCreatorAddress,
-  abi: tokenBridgeCreatorABI,
-} as const;
