+++ src/getBatchPosters.ts
@@ -8,20 +8,24 @@ import {
   getAbiItem,
   getFunctionSelector,
 } from 'viem';
-import { rollupCreator, upgradeExecutor } from './contracts';
-import { safeL2ABI, sequencerInboxABI } from './abi';
+
+import { rollupCreatorABI } from './contracts/RollupCreator';
+import { sequencerInboxABI } from './contracts/SequencerInbox';
+import { upgradeExecutorABI } from './contracts/UpgradeExecutor';
+import { gnosisSafeL2ABI } from './contracts/GnosisSafeL2';
+
 import { createRollupFetchTransactionHash } from './createRollupFetchTransactionHash';
 
-const createRollupABI = getAbiItem({ abi: rollupCreator.abi, name: 'createRollup' });
+const createRollupABI = getAbiItem({ abi: rollupCreatorABI, name: 'createRollup' });
 const createRollupFunctionSelector = getFunctionSelector(createRollupABI);
 
 const setIsBatchPosterABI = getAbiItem({ abi: sequencerInboxABI, name: 'setIsBatchPoster' });
 const setIsBatchPosterFunctionSelector = getFunctionSelector(setIsBatchPosterABI);
 
-const executeCallABI = getAbiItem({ abi: upgradeExecutor.abi, name: 'executeCall' });
+const executeCallABI = getAbiItem({ abi: upgradeExecutorABI, name: 'executeCall' });
 const upgradeExecutorExecuteCallFunctionSelector = getFunctionSelector(executeCallABI);
 
-const execTransactionABI = getAbiItem({ abi: safeL2ABI, name: 'execTransaction' });
+const execTransactionABI = getAbiItem({ abi: gnosisSafeL2ABI, name: 'execTransaction' });
 const safeL2FunctionSelector = getFunctionSelector(execTransactionABI);
 
 const ownerFunctionCalledEventAbi = getAbiItem({
